+++ src/app/header/home/home.component.ts
@@ -1,5 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 
+import * as firebase from 'firebase/app';
+import 'firebase/auth';
+
+import { AuthService } from '../../services/auth.service';
+
 @Component({
   selector: 'app-home',
   templateUrl: './home.component.html',
@@ -7,9 +12,12 @@ import { Component, OnInit } from '@angular/core';
 })
 export class HomeComponent implements OnInit {
 
-  constructor() { }
+  constructor(
+    public authService: AuthService
+  ) { }
 
-  ngOnInit() {
+  ngOnInit(): void {
+    this.authService.watchAuthState();
   }
 
 }