+++ src/app/models/post.ts
@@ -3,5 +3,7 @@ export class Post {
   title: string;
   content: string;
   loveIts: number;
-  created_at: Date;
+  hateIts: number;
+  created_at: string;
+  photo: string;
 }