+++ src/app/posts/new-post/new-post.component.ts
@@ -1,5 +1,6 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { Subscription } from 'rxjs';
 import { Router } from '@angular/router';
 
 import { PostService } from '../../services/post.service';
@@ -11,10 +12,14 @@ import { Post } from '../../models/post';
   templateUrl: './new-post.component.html',
   styleUrls: ['./new-post.component.css']
 })
-export class NewPostComponent implements OnInit {
+export class NewPostComponent implements OnInit, OnDestroy {
 
   postForm: FormGroup;
   posts: Post[];
+  postsSubscription: Subscription;
+  fileIsUploading: boolean = false;
+  fileUrl: string;
+  fileUploaded: boolean = false;
 
   constructor(
     private postService: PostService,
@@ -24,6 +29,7 @@ export class NewPostComponent implements OnInit {
 
   ngOnInit() {
     this.initForm();
+    this.getPosts();
   }
 
   initForm(): void {
@@ -42,10 +48,38 @@ export class NewPostComponent implements OnInit {
       title: title,
       content: content,
       loveIts: 0,
-      created_at: new Date()
+      hateIts: 0,
+      created_at: Date(),
+      photo: this.fileUrl ? this.fileUrl : ''
     };
     this.postService.addPost(newPost);
     this.router.navigate(['/posts']);
+  }
+
+  onUploadFile(file: File) {
+    this.fileIsUploading = true;
+    this.postService.uploadFile(file).then(
+      (url: string) => {
+        this.fileUrl = url;
+        this.fileIsUploading = false;
+        this.fileUploaded = true;
+      }
+    );
+  }
+
+  detectFiles(event) {
+    this.onUploadFile(event.target.files[0]);
+  }
+
+  getPosts() {
+    this.postsSubscription = this.postService.postsSubject.subscribe((posts: Post[]) => {
+      this.posts = posts;
+    });
+    this.postService.getPosts();
+  }
+
+  ngOnDestroy() {
+    this.postsSubscription.unsubscribe();
   }
 
 }