+++ RELEASE_NOTES.md
@@ -1,6 +1,11 @@
+Version 1.2.0 (28/06/2018)
+-----------------------------
+* Implementation of firebase database to store posts
+* Implementation of firebase storage to store images
+
 Version 1.1.0 (24/06/2018)
 -----------------------------
-* Firebase authentication
+* Implementation of firebase authentication: email/password
 
 Version 1.0.0 (19/06/2018)
 -----------------------------